fix: Enqueue foundational `wp-edit-post` asset
This asset is paramount for a functioning editor, but is absent if not
explicitly enqueued, as it is a top-level, entry dependency.

Eventually, the endpoint should likely conditionally enqueue the
relevant asset depending upon the targeted entity, likely communicated
via a query parameter. E.g., editing a site should enqueue
`wp-edit-site`.

See:

- https://github.com/WordPress/wordpress-develop/blob/4115c99a784db8a787e2245300c83c9c8d336503/src/wp-includes/script-loader.php#L302-L304
- https://github.com/WordPress/gutenberg/blob/f8a19eb08945a9fd26d5bda49b1e2c94857ea827/lib/client-assets.php#L215-L217